requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-compat==1.0.15
django4-background-tasks==1.2.9
et-xmlfile==1.1.0
numpy==1.24.4
openpyxl==3.1.4
pandas==2.0.3
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
